requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
drf-spectacular-sidecar==2023.7.1
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.6.1
Pillow==10.0.0
PyJWT==2.7.0
python-magic==0.4.27
pytz==2023.3
PyYAML==6.0.1
referencing==0.29.3
rpds-py==0.8.11
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
